requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
requests==2.32.3
python-dateutil==2.9.0.post0
dateparser==1.2.0
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
streamlit==1.37.1
google-generativeai==0.7.2
google-cloud-speech==2.27.0
